refactor selectors/basic-event-filters to use new Event model;
